[TdP] F5.1 and F6.1

+++ backend/src/main/java/pt/ulisboa/tecnico/socialsoftware/tutor/tournament/TournamentService.java
@@ -6,7 +6,7 @@ import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
-import pt.ulisboa.tecnico.socialsoftware.tutor.answer.AnswerService;
+
 import pt.ulisboa.tecnico.socialsoftware.tutor.answer.domain.QuizAnswer;
 import pt.ulisboa.tecnico.socialsoftware.tutor.answer.repository.QuizAnswerRepository;
 import pt.ulisboa.tecnico.socialsoftware.tutor.config.DateHandler;
@@ -27,6 +27,7 @@ import pt.ulisboa.tecnico.socialsoftware.tutor.quiz.domain.QuizQuestion;
 import pt.ulisboa.tecnico.socialsoftware.tutor.quiz.repository.QuizQuestionRepository;
 import pt.ulisboa.tecnico.socialsoftware.tutor.quiz.repository.QuizRepository;
 
+import pt.ulisboa.tecnico.socialsoftware.tutor.statement.dto.StatementQuizDto;
 import pt.ulisboa.tecnico.socialsoftware.tutor.tournament.domain.Tournament;
 import pt.ulisboa.tecnico.socialsoftware.tutor.tournament.dto.TournamentDto;
 import pt.ulisboa.tecnico.socialsoftware.tutor.tournament.repository.TournamentRepository;
@@ -34,13 +35,10 @@ import pt.ulisboa.tecnico.socialsoftware.tutor.user.User;
 import pt.ulisboa.tecnico.socialsoftware.tutor.user.UserRepository;
 
 import java.sql.SQLException;
-
+import java.util.*;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 
-import java.util.HashSet;
-import java.util.List;
-
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import static pt.ulisboa.tecnico.socialsoftware.tutor.exceptions.ErrorMessage.*;
@@ -65,13 +63,13 @@ public class TournamentService{
     private QuizRepository quizRepository;
 
     @Autowired
-    private QuizAnswerRepository quizAnswerRepository;
-
-    @Autowired
     private QuestionRepository questionRepository;
 
     @Autowired
     private QuizQuestionRepository quizQuestionRepository;
+
+    @Autowired
+    private QuizAnswerRepository quizAnswerRepository;
 
 
     @Retryable(
@@ -115,8 +113,25 @@ public class TournamentService{
             backoff = @Backoff(delay = 5000))
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     public List<TournamentDto> showAllOpenTournaments(int executionId){
-            return tournamentRepository.findOpen(executionId).stream().map(TournamentDto::new).collect(Collectors.toList());
+        return tournamentRepository.findAvailable(executionId).stream()
+                    .filter(tournament -> DateHandler.now().isBefore(tournament.getConclusionDate()))
+                    .filter(tournament -> tournament.getStartingDate().isBefore(DateHandler.now()))
+                    .map(TournamentDto::new).collect(Collectors.toList());
     }
+
+    @Retryable(
+            value = { SQLException.class },
+            backoff = @Backoff(delay = 5000))
+    @Transactional(isolation = Isolation.REPEATABLE_READ)
+    public List<TournamentDto> getDashBoardTournaments(int userId){
+        return tournamentRepository.findParticipated(userId).stream()
+                .filter(tournament -> DateHandler.now().isAfter(tournament.getConclusionDate()))
+                .map(TournamentDto::new).collect(Collectors.toList());
+
+
+    }
+
+
 
     @Retryable(
             value = { SQLException.class },
@@ -126,22 +141,8 @@ public class TournamentService{
 
         Tournament tournament = tournamentRepository.findById(tournamentId).orElseThrow(() -> new TutorException(TOURNAMENT_NOT_FOUND, tournamentId));
 
-        tournament.removeTopicList();
-        tournament.removeUsers();
-        tournament.removeCourseExecution();
-
-
-        Quiz quiz = tournament.getQuiz();
-        if (quiz != null){
-            quiz.remove();
-
-            Set<QuizQuestion> quizQuestions = new HashSet<>(quiz.getQuizQuestions());
-
-            quizQuestions.forEach(QuizQuestion::remove);
-            quizQuestions.forEach(quizQuestion -> quizQuestionRepository.delete(quizQuestion));
-
-            quizRepository.delete(quiz);
-            tournament.setQuiz(null);
+        if (tournament.getStartingDate().isBefore(DateHandler.now())){
+            throw new TutorException(TOURNAMENT_IS_OPEN, tournamentId);
         }
 
         tournamentRepository.delete(tournament);
@@ -155,7 +156,6 @@ public class TournamentService{
         return tournamentRepository.findById(tournamentId)
                 .map(TournamentDto::new)
                 .orElseThrow(() -> new TutorException(TOURNAMENT_NOT_FOUND, tournamentId));
-
     }
 
 
@@ -172,14 +172,14 @@ public class TournamentService{
         tournament.addStudent(user);
         user.getTournamentsEnrolled().add(tournament);
 
-        if (tournament.getStudentList().size() == 2 && tournament.getQuiz() == null){
-            generateTournamentQuiz(tournament, user);
+        if (tournament.getStudentList().size() == 2 && tournament.getQuiz() == null && tournament.isOpen()){
+            createTournamentQuiz(tournament);
         }
 
         return new TournamentDto(tournament);
     }
 
-    public void generateTournamentQuiz(Tournament tournament, User user){
+    public void createTournamentQuiz(Tournament tournament){
         Quiz quiz = new Quiz();
         quiz.setType(Quiz.QuizType.GENERATED.toString());
         quiz.setCreationDate(DateHandler.now());
@@ -190,28 +190,14 @@ public class TournamentService{
         List<Question> questions = questionRepository.findAvailableQuestions(courseExecution.getCourse().getId());
         // Get only questions with the correct topics
         List<Question> availableQuestions =  parseQuestions(questions, tournament);
-
-        /*
-        if(quizDetails.getAssessment() != null) {
-            availableQuestions = filterByAssessment(availableQuestions, quizDetails);
-        }
-        // TODO else use default assessment
-        */
-
-
         if (availableQuestions.size() < tournament.getNumberOfQuestions()) {
             throw new TutorException(NOT_ENOUGH_QUESTIONS);
         }
-
-        // ???
-        availableQuestions = user.filterQuestionsByStudentModel(tournament.getNumberOfQuestions(), availableQuestions);
 
         quiz.generate(availableQuestions);
         quiz.setTitle(tournament.getTitle() + " (Quiz)");
         quiz.setAvailableDate(tournament.getStartingDate());
         quiz.setConclusionDate(tournament.getConclusionDate());
-
-
         quiz.setCourseExecution(courseExecution);
         courseExecution.addQuiz(quiz);
 
@@ -225,7 +211,10 @@ public class TournamentService{
         for (Question q : questions){
             for (Topic t : q.getTopics()){
                 if (topics.contains(t)){
-                    availableQuestions.add(q);
+                    if (q.getImage() == null) // Images bug locally
+                        availableQuestions.add(q);
+                    if (availableQuestions.size()==tournament.getNumberOfQuestions())
+                        return availableQuestions;
                     continue;
                 }
             }
@@ -233,31 +222,6 @@ public class TournamentService{
 
         return availableQuestions;
     }
-
-    /*
-    public List<Question> parseQuestions(List<Question> questions, Tournament tournament){
-        List<Topic> topics = tournament.getTopicList();
-        List<Question> availableQuestions = new ArrayList<>();
-        boolean next = false;
-        for (Question q : questions){
-            for (Topic t : q.getTopics()){
-                for (Topic tt : topics) {
-                    if (t.getId() == tt.getId()) {
-                        availableQuestions.add(q);
-                        next = true;
-                        break;
-                    }
-                }
-                if (next){
-                    next = false;
-                    break;
-                }
-            }
-        }
-
-        return availableQuestions;
-    }
-    */
 
 
     @Retryable(
@@ -274,5 +238,48 @@ public class TournamentService{
                 .map(Tournament::getCourseExecution)
                 .map(CourseDto::new)
                 .orElseThrow(() -> new TutorException(TOURNAMENT_NOT_FOUND, tournamentId));
+    }
+
+    @Retryable(
+            value = { SQLException.class },
+            backoff = @Backoff(delay = 5000))
+    @Transactional(isolation = Isolation.REPEATABLE_READ)
+    public StatementQuizDto startTournament(int userId, int tournamentId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new TutorException(USER_NOT_FOUND, userId));
+        Tournament tournament = tournamentRepository.findById(tournamentId).orElseThrow(() -> new TutorException(TOURNAMENT_NOT_FOUND, tournamentId));
+        Quiz quiz = tournament.getQuiz();
+        LocalDateTime now = LocalDateTime.now();
+
+        // Check if quiz can be generated
+        if (tournament.getStudentList().size() < 2)
+            throw new TutorException(QUIZ_CANT_BE_GENERATED);
+
+        // Generate quiz if it hasn't been generated already
+        if (tournament.getQuiz() == null)
+            createTournamentQuiz(tournament);
+
+        // Check if user in enrolled
+        if (!tournament.getStudentList().contains(user))
+            throw new TutorException(STUDENT_NOT_ENROLLED);
+
+        // Check if tournament hasn't started
+        if (tournament.getStartingDate().isAfter(now))
+            throw new TutorException(TOURNAMENT_NOT_OPEN);
+
+        // Check if tournament is already finished
+        if (tournament.getConclusionDate().isBefore(now))
+            throw new TutorException(TOURNAMENT_IS_FINISHED);
+
+        // Check if user has already participated in the tournament
+        if (tournament.getAnsweredList().contains(user))
+            throw new TutorException(STUDENT_ALREADY_PARTICIPATED);
+
+        tournament.getAnsweredList().add(user);
+
+        QuizAnswer quizAnswer = new QuizAnswer(user, quiz);
+        quizAnswerRepository.save(quizAnswer);
+        StatementQuizDto statementQuizDto = new StatementQuizDto(quizAnswer);
+
+        return statementQuizDto;
     }
 }
